fix: 글쓰기 & 수정페이지 - 커버 이미지 업로드 예외 처리
