+++ src/components/post/index.ts
@@ -5,29 +5,6 @@ import PostCategories from "./PostCategories";
 import PostTravelStatus from "./PostTravelStatus";
 import PostMobileCourse from "./PostMobileCourse";
 
-const galleryLists = [
-  "https://cdn.pixabay.com/photo/2020/05/21/11/37/road-5200366_1280.jpg",
-  "https://cdn.pixabay.com/photo/2017/06/20/12/07/castle-2422860_1280.jpg",
-  "https://cdn.pixabay.com/photo/2020/02/09/00/21/jusang-joint-4831628_1280.jpg",
-  "https://cdn.pixabay.com/photo/2019/09/26/23/00/han-river-4507176_1280.jpg",
-  "https://cdn.pixabay.com/photo/2020/07/06/22/55/cornus-5378575_1280.jpg",
-  "https://cdn.pixabay.com/photo/2022/08/31/15/07/seoul-7423593_1280.jpg",
-  "https://cdn.pixabay.com/photo/2018/04/11/00/04/sea-3309231_1280.jpg",
-  "https://cdn.pixabay.com/photo/2019/09/17/02/20/jeju-4482313_1280.jpg",
-  "https://cdn.pixabay.com/photo/2017/01/17/15/18/haeundae-beach-1987193_1280.jpg",
-  "https://cdn.pixabay.com/photo/2022/04/04/13/54/city-7111380_1280.jpg",
-  "https://cdn.pixabay.com/photo/2015/09/10/14/11/jeju-934479_1280.jpg",
-  "https://cdn.pixabay.com/photo/2017/09/14/06/33/jeju-2748098_1280.jpg",
-  "https://cdn.pixabay.com/photo/2020/05/05/07/52/republic-of-korea-5131925_1280.jpg",
-  "https://cdn.pixabay.com/photo/2022/02/08/06/18/bird-7000837_1280.jpg",
-  "https://cdn.pixabay.com/photo/2021/09/01/14/57/gamcheon-culture-village-6591589_1280.jpg",
-  "https://cdn.pixabay.com/photo/2018/09/02/07/07/south-korea-3648252_1280.jpg",
-  "https://cdn.pixabay.com/photo/2017/09/26/17/06/and-thu-2789325_1280.jpg",
-  "https://cdn.pixabay.com/photo/2018/08/01/03/40/autumn-leaves-3576458_1280.jpg",
-  "https://cdn.pixabay.com/photo/2020/01/29/09/36/snow-4801975_1280.jpg",
-  "https://cdn.pixabay.com/photo/2015/10/17/15/54/seoul-olympic-park-992727_1280.jpg",
-];
-
 export {
   PostCourse,
   PostHashTag,
@@ -35,5 +12,4 @@ export {
   PostCategories,
   PostTravelStatus,
   PostMobileCourse,
-  galleryLists,
 };
